refactor(platforms): move manifest authority below adapters
